+++ ffi/capi/src/segmenter_line.rs
@@ -44,6 +44,7 @@ pub mod ffi {
     }
 
     #[diplomat::rust_link(icu::segmenter::options::LineBreakOptions, Struct)]
+    #[diplomat::rust_link(icu::segmenter::options::LineBreakOptions::default, FnInStruct, hidden)]
     #[diplomat::attr(supports = non_exhaustive_structs, rename = "LineBreakOptions")]
     #[diplomat::attr(kotlin, disable)] // option support (https://github.com/rust-diplomat/diplomat/issues/989)
     pub struct LineBreakOptionsV2 {
@@ -103,6 +104,21 @@ pub mod ffi {
         pub fn create_dictionary() -> Box<LineSegmenter> {
             Box::new(LineSegmenter(
                 icu_segmenter::LineSegmenter::new_dictionary(Default::default()).static_to_owned(),
+            ))
+        }
+
+        /// Construct a [`LineSegmenter`] with default options (no locale-based tailoring) and no support for scripts requiring complex context dependent line breaks
+        /// (Burmese, Khmer, Lao, and Thai), using compiled data
+        #[diplomat::rust_link(
+            icu::segmenter::LineSegmenter::new_for_non_complex_scripts,
+            FnInStruct
+        )]
+        #[diplomat::attr(auto, named_constructor = "for_non_complex_scripts")]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_for_non_complex_scripts() -> Box<LineSegmenter> {
+            Box::new(LineSegmenter(
+                icu_segmenter::LineSegmenter::new_for_non_complex_scripts(Default::default())
+                    .static_to_owned(),
             ))
         }
 
@@ -216,7 +232,6 @@ pub mod ffi {
         #[diplomat::attr(all(not(supports = non_exhaustive_structs), supports = fallible_constructors, supports = named_constructors), named_constructor = "dictionary_with_options_v2_and_provider")]
         #[cfg(feature = "buffer_provider")]
         #[diplomat::attr(kotlin, disable)] // option support (https://github.com/rust-diplomat/diplomat/issues/989)
-
         pub fn create_dictionary_with_options_v2_and_provider(
             provider: &DataProvider,
             content_locale: Option<&Locale>,
@@ -227,6 +242,55 @@ pub mod ffi {
 
             Ok(Box::new(LineSegmenter(
                 icu_segmenter::LineSegmenter::try_new_dictionary_with_buffer_provider(
+                    provider.get()?,
+                    options,
+                )?,
+            )))
+        }
+        /// Construct a [`LineSegmenter`] with custom options and no support for scripts requiring complex context dependent line breaks
+        /// (Burmese, Khmer, Lao, and Thai), using compiled data.
+        #[diplomat::rust_link(
+            icu::segmenter::LineSegmenter::new_for_non_complex_scripts,
+            FnInStruct
+        )]
+        #[diplomat::attr(supports = non_exhaustive_structs, rename = "for_non_complex_scripts_with_options")]
+        #[diplomat::attr(all(supports = non_exhaustive_structs, supports = named_constructors), named_constructor = "for_non_complex_scripts_with_options")]
+        #[diplomat::attr(all(not(supports = non_exhaustive_structs), supports = named_constructors), named_constructor = "for_non_complex_scripts_with_options_v2")]
+        #[cfg(feature = "compiled_data")]
+        #[diplomat::attr(kotlin, disable)] // option support (https://github.com/rust-diplomat/diplomat/issues/989)
+        pub fn create_for_non_complex_scripts_with_options_v2(
+            content_locale: Option<&Locale>,
+            options: LineBreakOptionsV2,
+        ) -> Box<LineSegmenter> {
+            let mut options: LineBreakOptions = options.into();
+            options.content_locale = content_locale.map(|c| &c.0.id);
+
+            Box::new(LineSegmenter(
+                icu_segmenter::LineSegmenter::new_for_non_complex_scripts(options)
+                    .static_to_owned(),
+            ))
+        }
+        /// Construct a [`LineSegmenter`] with custom options and no support for complex languages
+        /// (Burmese, Khmer, Lao, and Thai), using a particular data source.
+        #[diplomat::rust_link(
+            icu::segmenter::LineSegmenter::new_for_non_complex_scripts,
+            FnInStruct
+        )]
+        #[diplomat::attr(supports = non_exhaustive_structs, rename = "for_non_complex_scripts_with_options_and_provider")]
+        #[diplomat::attr(all(supports = non_exhaustive_structs, supports = fallible_constructors, supports = named_constructors), named_constructor = "for_non_complex_scripts_with_options_and_provider")]
+        #[diplomat::attr(all(not(supports = non_exhaustive_structs), supports = fallible_constructors, supports = named_constructors), named_constructor = "for_non_complex_scripts_with_options_v2_and_provider")]
+        #[cfg(feature = "buffer_provider")]
+        #[diplomat::attr(kotlin, disable)] // option support (https://github.com/rust-diplomat/diplomat/issues/989)
+        pub fn create_for_non_complex_scripts_with_options_v2_and_provider(
+            provider: &DataProvider,
+            content_locale: Option<&Locale>,
+            options: LineBreakOptionsV2,
+        ) -> Result<Box<LineSegmenter>, DataError> {
+            let mut options: LineBreakOptions = options.into();
+            options.content_locale = content_locale.map(|c| &c.0.id);
+
+            Ok(Box::new(LineSegmenter(
+                icu_segmenter::LineSegmenter::try_new_for_non_complex_scripts_with_buffer_provider(
                     provider.get()?,
                     options,
                 )?,