+++ ffi/mvn/src/main/kotlin/src/main/kotlin/org/unicode/icu4x/WordSegmenter.kt
@@ -16,6 +16,9 @@ internal interface WordSegmenterLib: Library {
     fun icu4x_WordSegmenter_create_dictionary_mv1(): Pointer
     fun icu4x_WordSegmenter_create_dictionary_with_content_locale_mv1(locale: Pointer): ResultPointerInt
     fun icu4x_WordSegmenter_create_dictionary_with_content_locale_and_provider_mv1(provider: Pointer, locale: Pointer): ResultPointerInt
+    fun icu4x_WordSegmenter_create_for_non_complex_scripts_mv1(): Pointer
+    fun icu4x_WordSegmenter_create_for_non_complex_scripts_with_content_locale_mv1(locale: Pointer): ResultPointerInt
+    fun icu4x_WordSegmenter_create_for_non_complex_scripts_with_content_locale_and_provider_mv1(provider: Pointer, locale: Pointer): ResultPointerInt
     fun icu4x_WordSegmenter_segment_utf16_mv1(handle: Pointer, input: Slice): Pointer
 }
 /** An ICU4X word-break segmenter, capable of finding word breakpoints in strings.
@@ -40,7 +43,7 @@ class WordSegmenter internal constructor (
         internal val lib: WordSegmenterLib = Native.load("icu4x", libClass)
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with automatically selecting the best available LSTM
+        /** Construct a [WordSegmenter] with automatically selecting the best available LSTM
         *or dictionary payload data, using compiled data. This does not assume any content locale.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and LSTM for Burmese,
@@ -59,7 +62,7 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with automatically selecting the best available LSTM
+        /** Construct a [WordSegmenter] with automatically selecting the best available LSTM
         *or dictionary payload data, using compiled data.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and LSTM for Burmese,
@@ -82,7 +85,7 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with automatically selecting the best available LSTM
+        /** Construct a [WordSegmenter] with automatically selecting the best available LSTM
         *or dictionary payload data, using a particular data source.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and LSTM for Burmese,
@@ -105,8 +108,8 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with LSTM payload data for Burmese, Khmer, Lao, and
-        *Thai, using compiled data.  This does not assume any content locale.
+        /** Construct a [WordSegmenter] with LSTM payload data for Burmese, Khmer, Lao, and
+        *Thai, using compiled data. This does not assume any content locale.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and LSTM for Burmese,
         *Khmer, Lao, and Thai.
@@ -124,7 +127,7 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with LSTM payload data for Burmese, Khmer, Lao, and
+        /** Construct a [WordSegmenter] with LSTM payload data for Burmese, Khmer, Lao, and
         *Thai, using compiled data.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and LSTM for Burmese,
@@ -147,7 +150,7 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with LSTM payload data for Burmese, Khmer, Lao, and
+        /** Construct a [WordSegmenter] with LSTM payload data for Burmese, Khmer, Lao, and
         *Thai, using a particular data source.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and LSTM for Burmese,
@@ -170,8 +173,8 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with with dictionary payload data for Chinese, Japanese,
-        *Burmese, Khmer, Lao, and Thai, using compiled data.  This does not assume any content locale.
+        /** Construct a [WordSegmenter] with dictionary payload data for Chinese, Japanese,
+        *Burmese, Khmer, Lao, and Thai, using compiled data. This does not assume any content locale.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and dictionary for Burmese,
         *Khmer, Lao, and Thai.
@@ -189,7 +192,7 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with dictionary payload data for Chinese, Japanese,
+        /** Construct a [WordSegmenter] with dictionary payload data for Chinese, Japanese,
         *Burmese, Khmer, Lao, and Thai, using compiled data.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and dictionary for Burmese,
@@ -212,7 +215,7 @@ class WordSegmenter internal constructor (
         }
         @JvmStatic
         
-        /** Construct an [WordSegmenter] with dictionary payload data for Chinese, Japanese,
+        /** Construct a [WordSegmenter] with dictionary payload data for Chinese, Japanese,
         *Burmese, Khmer, Lao, and Thai, using a particular data source.
         *
         *Note: currently, it uses dictionary for Chinese and Japanese, and dictionary for Burmese,
@@ -223,6 +226,62 @@ class WordSegmenter internal constructor (
         fun createDictionaryWithContentLocaleAndProvider(provider: DataProvider, locale: Locale): Result<WordSegmenter> {
             
             val returnVal = lib.icu4x_WordSegmenter_create_dictionary_with_content_locale_and_provider_mv1(provider.handle, locale.handle);
+            if (returnVal.isOk == 1.toByte()) {
+                val selfEdges: List<Any> = listOf()
+                val handle = returnVal.union.ok 
+                val returnOpaque = WordSegmenter(handle, selfEdges)
+                CLEANER.register(returnOpaque, WordSegmenter.WordSegmenterCleaner(handle, WordSegmenter.lib));
+                return returnOpaque.ok()
+            } else {
+                return DataErrorError(DataError.fromNative(returnVal.union.err)).err()
+            }
+        }
+        @JvmStatic
+        
+        /** Construct a [WordSegmenter] with no support for scripts requiring complex context dependent word breaks (Chinese, Japanese,
+        *Burmese, Khmer, Lao, and Thai), using compiled data. This does not assume any content locale.
+        *
+        *See the [Rust documentation for `new_for_non_complex_scripts`](https://docs.rs/icu/2.1.1/icu/segmenter/struct.WordSegmenter.html#method.new_for_non_complex_scripts) for more information.
+        */
+        fun createForNonComplexScripts(): WordSegmenter {
+            
+            val returnVal = lib.icu4x_WordSegmenter_create_for_non_complex_scripts_mv1();
+            val selfEdges: List<Any> = listOf()
+            val handle = returnVal 
+            val returnOpaque = WordSegmenter(handle, selfEdges)
+            CLEANER.register(returnOpaque, WordSegmenter.WordSegmenterCleaner(handle, WordSegmenter.lib));
+            return returnOpaque
+        }
+        @JvmStatic
+        
+        /** Construct a [WordSegmenter] with no support for scripts requiring complex context dependent word breaks (Chinese, Japanese,
+        *Burmese, Khmer, Lao, and Thai), using compiled data.
+        *
+        *See the [Rust documentation for `try_new_for_non_complex_scripts`](https://docs.rs/icu/2.1.1/icu/segmenter/struct.WordSegmenter.html#method.try_new_for_non_complex_scripts) for more information.
+        */
+        fun createForNonComplexScriptsWithContentLocale(locale: Locale): Result<WordSegmenter> {
+            
+            val returnVal = lib.icu4x_WordSegmenter_create_for_non_complex_scripts_with_content_locale_mv1(locale.handle);
+            if (returnVal.isOk == 1.toByte()) {
+                val selfEdges: List<Any> = listOf()
+                val handle = returnVal.union.ok 
+                val returnOpaque = WordSegmenter(handle, selfEdges)
+                CLEANER.register(returnOpaque, WordSegmenter.WordSegmenterCleaner(handle, WordSegmenter.lib));
+                return returnOpaque.ok()
+            } else {
+                return DataErrorError(DataError.fromNative(returnVal.union.err)).err()
+            }
+        }
+        @JvmStatic
+        
+        /** Construct a [WordSegmenter] with no support for scripts requiring complex context dependent word breaks (Chinese, Japanese,
+        *Burmese, Khmer, Lao, and Thai), using a particular data source.
+        *
+        *See the [Rust documentation for `try_new_for_non_complex_scripts`](https://docs.rs/icu/2.1.1/icu/segmenter/struct.WordSegmenter.html#method.try_new_for_non_complex_scripts) for more information.
+        */
+        fun createForNonComplexScriptsWithContentLocaleAndProvider(provider: DataProvider, locale: Locale): Result<WordSegmenter> {
+            
+            val returnVal = lib.icu4x_WordSegmenter_create_for_non_complex_scripts_with_content_locale_and_provider_mv1(provider.handle, locale.handle);
             if (returnVal.isOk == 1.toByte()) {
                 val selfEdges: List<Any> = listOf()
                 val handle = returnVal.union.ok